Main/requirements.txt
numpy
matplotlib
opencv-python
pytesseract
pillow
scikit-learn
tensorflow
tk